requirements.txt
streamlit==1.25.0
pandas==1.5.3
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.2
scikit-learn==1.3.0
matplotlib==3.7.1
seaborn==0.12.2
python-dotenv==1.0.0
openai==0.27.4
nltk==3.9.0
textblob==0.17.0
tqdm==4.65.0
